requirements.txt
pymongo
openai
certifi
python-dotenv
